:recycle: Centralize and reuse test fixtures with conftest.py
- Moved shared test fixtures (`user`, `staff_user`,
`authenticated_client`, etc.) to a dedicated `conftest.py` file.

- Removed duplicate definitions of these fixtures from `test_views.py`.

- Ensured consistent import and reuse of fixtures and utility functions across tests.
